Ajout de la gestion de creation d'album cote front avec responsive

+++ back-end/controllers/albumController.js
@@ -64,30 +64,39 @@ export const updateCover = async (req, res) => {
   let file;
   try {
     const { id } = req.params;
-    file = req.file.path;
+    const { title } = req.body;
+    let updateFields = {};
 
-    const result = await cloudinary.v2.uploader.upload(file, {
-      folder: "album/covers",
+    // Si un fichier est envoyé, upload sur Cloudinary
+    if (req.file) {
+      file = req.file.path;
+      const result = await cloudinary.v2.uploader.upload(file, {
+        folder: "album/covers",
+      });
+      updateFields.coverImage = result.secure_url;
+      // Supprime le fichier temporaire après upload
+      fs.unlink(file, (err) => {
+        if (err) console.error("Erreur suppression fichier temporaire:", err);
+      });
+    }
+    // Si un titre est envoyé, on le met à jour
+    if (title) {
+      updateFields.title = title;
+    }
+    if (Object.keys(updateFields).length === 0) {
+      return res
+        .status(400)
+        .json({ message: "Aucune donnée à mettre à jour." });
+    }
+    const updateAlbum = await Album.findByIdAndUpdate(id, updateFields, {
+      new: true,
     });
-
-    // Supprime le fichier temporaire après upload
-    fs.unlink(file, (err) => {
-      if (err) console.error("Erreur suppression fichier temporaire:", err);
-    });
-
-    const updateAlbum = await Album.findByIdAndUpdate(
-      id,
-      { coverImage: result.secure_url },
-      { new: true }
-    );
-
     res.status(200).json(updateAlbum);
   } catch (error) {
-    // Supprime le fichier temporaire en cas d'erreur
     if (file) fs.unlink(file, () => {});
     console.error(error);
     res.status(500).json({
-      message: "Erreur lors de la modification de l'image de couverture",
+      message: "Erreur lors de la modification de l'album",
     });
   }
 };
